# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CreatedIdResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CreatedIdResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CreatedIdResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"StaffResetPasswordBody": {
				"properties": {
					"newPassword": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffUpdateStatusBody": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "inactive"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CustomerResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"first_name",
					"last_name",
					"email",
					"phone",
					"address",
					"avatar",
					"status",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CustomerResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CustomerResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CustomerListResponse": {
				"properties": {
					"customers": {
						"items": {
							"$ref": "#/components/schemas/CustomerResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"customers",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CustomerListResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CustomerListResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCustomerBody": {
				"properties": {
					"username": {
						"type": "string",
						"example": "user1"
					},
					"first_name": {
						"type": "string",
						"example": "Tran"
					},
					"last_name": {
						"type": "string",
						"example": "Tai"
					},
					"email": {
						"type": "string",
						"example": "user1@example.com"
					},
					"password": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"username",
					"first_name",
					"last_name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCustomerBody": {
				"properties": {
					"first_name": {
						"type": "string",
						"example": "Tran"
					},
					"last_name": {
						"type": "string",
						"example": "Tai"
					},
					"email": {
						"type": "string",
						"example": "user1@example.com"
					},
					"phone": {
						"type": "string",
						"example": "0778946513"
					},
					"address": {
						"type": "string",
						"example": "51 Duong A, TPHCM"
					},
					"avatar": {
						"type": "string",
						"example": "https://example.com/avatar.png"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCustomerStatusBody": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "inactive"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminResetPasswordBody": {
				"properties": {
					"new_password": {
						"type": "string",
						"example": "NewPass@123"
					}
				},
				"required": [
					"new_password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BMD Training API",
		"version": "1.0.0",
		"description": "API documentation for BMD Training backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/admin/staff/reset-password": {
			"put": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Reset password for a staff member by ID",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffResetPasswordBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff/status": {
			"put": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update status of a staff member by ID",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffUpdateStatusBody"
							}
						}
					}
				}
			}
		},
...
		"/api/admin/customers": {
			"get": {
				"operationId": "GetAllCustomers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CustomerListResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get all customers",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CreatedIdResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Create a new customer account",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCustomerBody"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update a customer",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCustomerBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Delete a customer",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/customers/{customerId}": {
			"get": {
				"operationId": "GetCustomerById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get a customer by ID",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/customers/status": {
			"put": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update customer status",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCustomerStatusBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/customers/restore": {
			"put": {
				"operationId": "RestoreCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Restore a deleted customer",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/customers/reset-password": {
			"put": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Reset a customer's password (admin only)",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminResetPasswordBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}